aspera/pyproject.toml
[project]
name = "aspera"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.6",
    "greenlet>=3.1.1",
    "llama-index-llms-dashscope>=0.3.0",
    "llama-index>=0.12.8",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy[asyncio]>=2.0.36",
    "llama-index-embeddings-dashscope>=0.3.0",
    "llama-index-vector-stores-postgres>=0.4.1",
    "llama-index-readers-web>=0.3.3",
    "arize-phoenix>=7.3.2",
    "arize-phoenix-otel>=0.6.1",
    "openinference-instrumentation-llama-index>=3.1.2",
    "trafilatura>=2.0.0",
]
